% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finsyncR.R
\docType{package}
\name{finsyncR}
\alias{finsyncR}
\title{finsyncR: An R Package to Access Cleaned and Compiled USGS and EPA Stream Data}
\description{
The \code{finsyncR} package provides the data and functions to generate
  site by taxa datasets for macroinvertebrates, fish, and algal samples.
  Additionally, the package can provide county-level pesticide use estimates.
}
