% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlgaeData.R
\name{getAlgaeData}
\alias{getAlgaeData}
\title{Access clean USGS Algae Dataset}
\usage{
getAlgaeData(
  algType = "peri",
  dataType = "abun",
  taxonLevel = "Species",
  program = "National Water Quality Assessment"
)
}
\arguments{
\item{algType}{Algae sample type, either \code{"peri"} or \code{"phyto"}.}

\item{dataType}{Output data type, either \code{"abun"} or \code{"occur"}.}

\item{taxonLevel}{Level of taxonomic resolution, must be one of:
\code{"AlgalGroup"}, \code{"Phylum"}, \code{"Class"}, \code{"Order"},
\code{"Family"}, \code{"Genus"}, or \code{"Species"}.}

\item{program}{The program name(s) that should be included in the output
dataset.}
}
\value{
A species by sample data frame with site, stream reach, and
  sample information.
}
\description{
Access clean USGS Algae Dataset
}
\details{
Note: Abundances for taxa from replicate samples taken from the
  same stream reach are averaged.

  \code{program} refers to the Local, regional, or national program project
  for which data were originally collected. Because the National Water
  Quality Assessment (NAWQA) contains the most standardized sampling methods,
  we recommend using only the NAWQA dataset (set as default) for density and
  abundance measures. If you want to use all datasets, set
  \code{program = "ALL"}. Otherwise, \code{program} can be set to
  \code{"National Water Quality Assessment"},
  \code{"Cooperative Water Program"},
  \code{"Collection of Basic Records"},
  \code{"Other Federal Agencies"}, or a combination of these using \code{c()}
  (for example, \code{program = c("National Water Quality Assessment",
  "Cooperative Water Program")} for both NAWQA and Cooperative Water Programs)
}
\examples{
\dontrun{
Periphyton <- getFishData(algType = "peri",
                          taxonLevel = "Family")

Phytoplankton <- getFishData(algType = "phyto",
                             taxonLevel = "Family")
}

}
