% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountyPest.R
\name{getCountyPest}
\alias{getCountyPest}
\title{Access clean USGS County Level Pesticide Estimates Dataset}
\usage{
getCountyPest(
  data,
  ePest = "low",
  lagTime = 0,
  lagType,
  pestLevel,
  pestLevelName
)
}
\arguments{
\item{data}{Input dataset which needs to include columns for
\code{"SiteNumber"} and \code{"CollectionYear"}.
\code{"SiteNumber"} must have \code{"USGS-"} prefix.}

\item{ePest}{Pesticide use estimate type. Must be \code{"low"} or
\code{"high"}.}

\item{lagTime}{numeric. The number of years over which the pesticide use
estimates should be summarized. 0 returns pesticide use estimates in
the given \code{"CollectionYear"}.}

\item{lagType}{Specifies whether to take a \code{"sum"} or a \code{"mean"}
over years. Can only be used if \code{lagTime} is greater than 0.}

\item{pestLevel}{Specifies whether to summarize pesticide use estimates
according to \code{"type"}, \code{"class"}, or \code{"compound"}.}

\item{pestLevelName}{Specifies the names of the pesticide types, classes, or
individual compounds. See details for usage.}
}
\value{
Function returns a site by year dataframe of county-level pesticides use estimates
  in kilograms.
}
\description{
Access clean USGS County Level Pesticide Estimates Dataset
}
\details{
Note: Pesticide classes and types taken from PAN Pesticide Database.
  There are 19 pesticides out of 483 that do not have a specific
  class or type assigned because they were not present in PAN Database.
  Some types were reassigned from PAN designations: plant growth regulators
  were designated as herbicides, and insect growth regulators were designated as
  insecticides. Fungicides were split into antibiotic fungicides, biologic fungicides,
  copper fungicides, mineral fungicides, and synthetic fungicides. To access the original PAN
  type designations see \code{Type1} in \code{StreamData:::.pest.info}.

  Not every county has an estimate for each of the 483 pesticides in every
  year. We have set \code{na.rm = TRUE} when means and sums are calculated,
  so we inherently assume use is 0 when not provided. Seed coats including
  many neonicotinoids are not included in pesticide use estimates. Pesticide use
  estimates are based on agriculture alone.

  See the following internal dataset for pesticide compounds, classes, and types
  \code{StreamData:::.pest.info}.
}
\examples{
\dontrun{
## Code to generate the sum of the county-level use estimates (high estimate)
## from site number "05276005" from 1997-1999 (2 year lag) for "insecticide",
## "herbicide", and "synthetic_fungicide".

  dat = data.frame(SiteNumber = "USGS-05276005",
                   CollectionYear = 1999)

  getCountyPest(data=dat,
  ePest = "high",
  lagType = "sum",
  lagTime = 2,
  pestLevel = "type",
  pestLevelName = c("insecticide", "herbicide", "synthetic_fungicide"))
  }

}
