% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFishData.R
\name{getFishData}
\alias{getFishData}
\title{Access clean USGS Fish Dataset}
\usage{
getFishData(
  dataType = "abun",
  taxonLevel = "Species",
  program = "National Water Quality Assessment",
  standardize = TRUE
)
}
\arguments{
\item{dataType}{Output data type, either \code{"abun"} or \code{"occur"}.}

\item{taxonLevel}{Level of taxonomic resolution, must be one of:
\code{"Superclass"}, \code{"Class"}, \code{"Subclass"}, \code{"Superorder"},
\code{"Order"}, \code{"Superfamily"}, \code{"Family"}, \code{"Subfamily"},
\code{"Genus"}, \code{"Species"}, or \code{"Subspecies"}.}

\item{program}{The program name(s) that should be included in the output
dataset. See \code{Details} below for more information.}

\item{standardize}{logical. Should abundance be standardized by unit effort (see
details)? \code{"TRUE"} or \code{"FALSE"}.}
}
\value{
A species by sample data frame with site, stream reach, and
  sample information.
}
\description{
Access clean USGS Fish Dataset
}
\details{
Note: To standardize fish abundance data, abundances were divided by
  the seconds shocked, number of seine hauls, etc. Then were divided by the
  stream length sampled. ~35% of the data have been removed,
  as these samples lacked the required standardization information.
  Therefore, if you are interested in occurrence (pres/abs) data only, then
  set \code{standardize = FALSE} and \code{dataType = "occur"}. Thus,
  be aware that setting \code{standardize = FALSE} will result in a larger
  dataset with additional samples/sites than when \code{standardize = TRUE}.

  \code{program} refers to the Local, regional, or national program project
  for which data were originally collected. Because the National Water
  Quality Assessment (NAWQA) contains the most standardized sampling methods,
  we recommend using only the NAWQA dataset (set as default) for density and
  abundance measures. If you want to use all datasets, set
  \code{program = "ALL"}. Otherwise, \code{program} can be set to
  \code{"National Water Quality Assessment"},
  \code{"Cooperative Water Program"},
  \code{"Collection of Basic Records"},
  \code{"Other Federal Agencies"}, or a combination of these using \code{c()}
  (for example, \code{program = c("National Water Quality Assessment",
  "Cooperative Water Program")} for both NAWQA and Cooperative Water
  Programs).

  The \code{getFishData()} function only outputs a community (taxa x site)
  matrix. We will include an extension of this function to return the
  individual-level information (length, width, deformities, etc.).
}
\examples{
\dontrun{
Fish <- getFishData(taxonLevel = "Family")

}


}
