% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvertData.R
\name{getInvertData}
\alias{getInvertData}
\title{Access clean USGS Macro-Invertebrate Dataset}
\usage{
getInvertData(
  dataType = "abun",
  taxonLevel = "Species",
  program = "National Water Quality Assessment",
  lifestage = FALSE,
  abunMeasure = "density"
)
}
\arguments{
\item{dataType}{Output data type, either \code{"abun"} or \code{"occur"}.}

\item{taxonLevel}{Level of taxonomic resolution, must be one of:
\code{"AlgalGroup"}, \code{"Phylum"}, \code{"Class"}, \code{"Order"},
\code{"Family"}, \code{"Genus"}, or \code{"Species"}.}

\item{program}{The program name(s) that should be included in the output
dataset. See \code{Details} below for more information.}

\item{lifestage}{logical. Should the output dataset should include lifestage information for
each individual? \code{TRUE} or \code{FALSE}.}

\item{abunMeasure}{Measure of abundance, either \code{"abundance"} or
\code{"density"}.}
}
\value{
A species by sample data frame with site, stream reach, and
  sample information.
}
\description{
Access clean USGS Macro-Invertebrate Dataset
}
\details{
Note: There are 81 sampling events (sampling location - collection
  date) with replicate samples (not duplicates). We have left these replicate
  samples in the dataset, because some replicates are different in the
  stream habitat sampled, which may be of interest for certain ecological
  questions.

  \code{program} refers to the Local, regional, or national program project
  for which data were originally collected. Because the National Water
  Quality Assessment (NAWQA) contains the most standardized sampling methods,
  we recommend using only the NAWQA dataset (set as default) for density and
  abundance measures. If you want to use all datasets, set
  \code{program = "ALL"}. Otherwise, \code{program} can be set to
  \code{"National Water Quality Assessment"},
  \code{"Cooperative Water Program"},
  \code{"Collection of Basic Records"},
  \code{"Other Federal Agencies"}, or a combination of these using \code{c()}
  (for example, \code{program = c("National Water Quality Assessment",
  "Cooperative Water Program")} for both NAWQA and Cooperative Water Programs)

  If \code{dataType = "abun"} AND \code{abunMeasure = "density"},
  ~0.5% of samples (22 samples) removed from the dataset, because area
  sampled was not provided for these samples.
}
\examples{
\dontrun{
Inverts <- getInvertData(taxonLevel = "Family")

}

}
