% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamData.R
\docType{package}
\name{StreamData}
\alias{StreamData}
\title{StreamData: An R Package to Access Cleaned and Compiled Stream Data for Powell Center Working Group}
\description{
The \code{StreamData} package provides the data and functions to generate
  the site by taxa datasets for macroinvertebrates, fish, and algal samples.
  Additionally, the package can provide county-level pesticide use estimates.
}
