% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finsyncR-package.R
\docType{package}
\name{finsyncR-package}
\alias{finsyncR}
\alias{finsyncR-package}
\title{finsyncR: An R Package to synchronize fish and invertebrate datasets from federal sources in the United States}
\description{
\code{finsyncR} (\strong{f}ish and \strong{in}vertebrate \strong{sync}hronizer in \strong{R}) provides
access to, manipulates, and harmonizes national-level aquatic biomonitoring
datasets in the United States, with a focus on fish
and macroinvertebrates sampled in rivers and streams. The package streamlines
the process of retrieving and working with these data, making the process of
utilizing these data  easier for researchers. The data resources used in this
package are from United States Environmental Protection Agency’s National
Aquatic Resources Surveys, namely the National River and Streams Assessment,
and United States Geological Survey’s BioData.

The package provides abundance (density) and occurrence data for fish and
macroinvertebrates at a variety of levels of taxonomic resolution. Alongside
community matrices, important data describing sampling methods are provided,
which are intended to be used as covariates in analyses. The package provides
flexibility for common data management issues for temporal biologic datasets.
Data management options include harmonization of macroinvertebrate taxonomy
across the two federal datasets and through time and the ability to drop
hybrid organisms altogether for fish. Samples, which often vary in the number
of individuals within a sample, can be rarefied to a specific count.
Additionally, fish samples can be standardized to catch per-unit effort and
to account for catchability differences between fish sampling methods. The
package can also match  sampling sites for fish and macroinvertebrates to
data from the National Land Cover Database.

Data from USGS BioData were obtained from Pete Ruhl (pmruhl@usgs.gov), and
data from the National Aquatic Resource Surveys are directly downloaded by
the package from the website: \link{https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys}.

Users should cite the aforementioned datasets as follows:
\itemize{
\item U.S. Geological Survey, 2020. BioData - Aquatic bioassessment data for the
Nation: U.S. Geological Survey database, accessed 17 December 2020,
at https://doi.org/10.5066/F77W698B
\item U.S. Environmental Protection Agency. 2006. National Aquatic Resource
Surveys. Wadeable Streams Assessment 2004 (data and metadata files).
Available from U.S. EPA web page:
https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys.
\item U.S. Environmental Protection Agency. 2016. National Aquatic Resource
Surveys. National Rivers and Streams Assessment 2008-2009 (data and metadata
files). Available from U.S. EPA web page:
https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys.
\item U.S. Environmental Protection Agency. 2020. National Aquatic Resource
Surveys. National Rivers and Streams Assessment 2013-2014 (data and metadata
files). Available from U.S. EPA web page:
https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys.
\item U.S. Environmental Protection Agency. 2022. National Aquatic Resource
Surveys. National Rivers and Streams Assessment 2018-2019 (data and metadata
files). Available from U.S. EPA web page:
https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys.
}

Metadata can be found in the "Metadata"
vignette: \link{vignette("Metadata", package = "finsyncR")}.

A compendium on the correct use of the finsyncR package and its output
datasets can be found in the "Getting Started"
vignette: \link{vignette("GettingStarted", package = "finsyncR")}.

\strong{Of note}: The approach provided in this package concerning the way the
USEPA National Rivers and Streams Assessment data are treated is not
necessarily the same as the approach used in the National Rivers and
Streams Assessment national reports.

\strong{Disclaimer}: The United States Environmental Protection Agency (EPA) GitHub
project code is provided on an "as is" basis and the user assumes
responsibility for its use. EPA has relinquished control of the information
and no longer has responsibility to protect the integrity , confidentiality,
or availability of the information. Any reference to specific commercial
products, processes, or services by service mark, trademark, manufacturer,
or otherwise, does not constitute or imply their endorsement, recommendation
or favoring by EPA. The EPA seal and logo shall not be used in any manner to
imply endorsement of any commercial product or activity by EPA or the United
States Government.
}
\author{
\strong{Maintainer}: Michael Mahon \email{mahon.michael@epa.gov}

Authors:
\itemize{
  \item Samantha Rumschlag \email{rumschlag.samantha@epa.gov}
  \item Devin Jones \email{devin.k.jones@gmail.com}
  \item Ethan Brown \email{ebrown23@nd.edu}
  \item Terry Brown \email{brown.terryn@epa.gov} [reviewer]
  \item Ryan Hill \email{hill.ryan@epa.gov}
  \item Stefan Kunz \email{stkunz@uni-landau.de}
}

}
\keyword{internal}
