% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFishData.R
\name{getFishData}
\alias{getFishData}
\title{Access clean USGS Fish Dataset}
\usage{
getFishData(
  dataType = "occur",
  taxonLevel = "Species",
  agency = c("USGS", "EPA"),
  standardize = "none",
  hybrids = TRUE
)
}
\arguments{
\item{dataType}{Output data type, either \code{"abun"} or \code{"occur"}.}

\item{taxonLevel}{Level of taxonomic resolution, must be one of:
\code{"Superclass"}, \code{"Class"}, \code{"Subclass"}, \code{"Superorder"},
\code{"Order"}, \code{"Superfamily"}, \code{"Family"}, \code{"Subfamily"},
\code{"Genus"}, \code{"Species"}, or \code{"Subspecies"}.}

\item{agency}{The agency name(s) (e.g., "USGS" and "EPA") that should be
included in the output dataset. As of now, "USGS" must be present in the
agency vector. See \code{Details} below for more information.}

\item{standardize}{Standardization method to use for calculating fish abundance matrices.
Default is \code{standardize = "none"}, which returns raw fish abundance values. Other options include
\code{standardize = "CPUE"}, which returns standardized abundances in Catch per Unit Effort.
An alternative standardization method is \code{standardize = "MGMS"}, which calculates
Multigear Mean Standardization (MGMS) values to account for catchability differences between
fish sampling methods (see 'details' or Gibson-Reinemer et al. (2014) for more info on MGMS).}

\item{hybrid}{logical. Should hybrid individuals be included in the output dataset? \code{TRUE} or \code{FALSE}.}
}
\value{
A species by sample data frame with site, stream reach, and
  sample information.
}
\description{
Access clean USGS Fish Dataset
}
\details{
Note: To standardize fish abundance data, abundances were divided by
  the seconds shocked, number of seine hauls, etc. Then were divided by the
  stream length sampled. ~35% of the data have been removed,
  as these samples lacked the required standardization information.
  Therefore, if you are interested in occurrence (pres/abs) data only, then
  set \code{dataType = "occur"}. Thus, be aware that setting \code{dataType = "occur"}
  will result in a larger dataset with additional samples/sites than when \code{dataType = "abun"}.

  To account for differences in efficacy between shocking, seine netting, and
  snorkeling, we included multigear mean standardization (\code{standardize = "MGMS"}) as a
  standardization method as an alternative to catch per unit effort (\code{standardize = "CPUE"}).
  To calculate MGMS, we first calculated the mean total CPUE (summation of CPUEs for all
  taxa at each site for each collection date) for each sampling method. Then we divided
  each CPUE value by the mean total CPUE for their respective sampling methods. See
  the supplement of Gibson-Reinemer et al. (2014) for an example of MGMS computation.

  Keep in mind that when setting \code{standardize = "CPUE"} will result in a larger dataset
  (more rows) than \code{standardize = "MGMS"} because CPUE data will have a row for each sampling
  method used at each time-location, whereas MGMS will have only one row for each time-location
  because the function is standardizing by sampling method in addition to standardizing by
  time and area sampled.

  \code{agency} refers to agency that collected the invertebrate samples. If
  you want to use data from both agencies, set \code{agency = c("USGS", "EPA")}.
  For the "USGS" dataset, this includes all programs with SampleMethodCodes
  of "BERW", "IRTH", "SWAMP", "EMAP", "CDPHE", and "PNAMP". If \code{agency}
  includes "EPA", samples from the EPA National Stream and River
  Assessment programs (2018-2019, 2013-2014, 2008-2009) will be included.
  Note that from these samples, only moving waters classified as "wadeable"
  are included and only samples that are "reach-wide" are included. Some
  information included in the USGS dataset are not included in the EPA
  datasets, and thus will appear as "NA". Similar to the USGS data, there
  were inherent taxonomic issues with the EPA data. As such, we have taken
  the same steps as described above under \code{taxonFix} to address these
  concerns. NOTE: As of now, \code{agency} must include "USGS". This will be
  fixed in future versions. NOTE2: As of now, when \code{agency = c("USGS", "EPA")},
  taxonLevel must be set to "Species" and dataType must be set to "occur".

  The \code{getFishData()} function only outputs a community (taxa x site)
  matrix. We will include an extension of this function to return the
  individual-level information (length, width, deformities, etc.).
}
\examples{
\dontrun{
Fish <- getFishData(taxonLevel = "Species")

}


}
\references{
Gibson-Reinemer DK, Ickes BS, Chick JH, 2014. Development and assessment of a new method for combining
catch per unit effort data from different fish sampling gears: multigear mean standardization (MGMS).
Can. J. Fish. Aquat. Sci. 74:8-14.
}
