% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvertData.R
\name{getInvertData}
\alias{getInvertData}
\title{Access and harmonize macroinvertebrate data}
\usage{
getInvertData(
  dataType = "occur",
  taxonLevel = "Genus",
  taxonFix = "lump",
  agency = c("USGS", "EPA"),
  lifestage = FALSE,
  rarefy = TRUE,
  sharedTaxa = FALSE,
  seed = 0,
  boatableStreams = FALSE
)
}
\arguments{
\item{dataType}{Output data type for the community matrix, either \code{"abun"}
(abundance) or \code{"occur"} (occurrence).}

\item{taxonLevel}{Level of taxonomic resolution for the community matrix. Input must be one of:
\code{"Phylum"}, \code{"Class"}, \code{"Order"},
\code{"Family"}, \code{"Genus"}, or \code{"Species"}.}

\item{taxonFix}{Option to account for changes in taxonomy across time, must be
one of: \code{"none"}, \code{"lump"}, \code{"remove"}. See \code{Details}
below for more information.}

\item{agency}{The agency name or names (e.g., "USGS" and "EPA") that are the
source of data for the output community matrix. See \code{Details} below for more information.}

\item{lifestage}{logical. For USGS data only, should the output dataset include
lifestage information for each individual?  \code{TRUE} or \code{FALSE}.}

\item{rarefy}{logical. Should samples be standardized by the number of individuals
identified within each sample? \code{TRUE} or \code{FALSE}. See \code{Details}
below for more information.}

\item{sharedTaxa}{logical. Should Genera be limited to those that appear in
both the EPA and USGS datasets? \code{TRUE} or \code{FALSE}. Must be set to
\code{FALSE} when only one agency is specified.}

\item{seed}{numeric. Set seed for \code{rarefy} to get consistent results with each new
iteration of the function. Value gets passed to \code{set.seed} internally.}

\item{boatableStreams}{logical. Should EPA boatable streams be included in the
output dataset?  \code{TRUE} or \code{FALSE}. Note: all USGS streams are wadable.
It is not advisable to include boatable streams when building a dataset including
both EPA and USGS data. Boatable EPA data and wadeable USGS data are not
considered comparable.}
}
\value{
A taxa by sample data frame with site, stream reach, and sample information.
}
\description{
This function generates an occurrence or abundance community matrix for
benthic macroinvertebrates sampled in rivers and streams.
}
\details{
\code{taxonFix} provides options to account for changes in taxonomy across time,
especially in instances in which species have been reorganized into new genera.
\code{taxonFix} operates on the genera level. \code{taxonFix = "none"}
makes no adjustment. \code{taxonFix = "lump"} prioritizes retaining observations
by giving a unified genera name to all species and genera that have been linked
through changes in taxonomy through time (e.g. genera1/genera2/genera3). Note:
of 98 problematic genera that exist throughout both datasets,
\code{taxonFix = "lump"} results in 13 "lumped" genera. Most new "lumped"
genera are fewer than five "old" genera joined. One "lumped" genera (within
Ephemeroptera) includes 70 unique genera names. So, it is not advisable to
examine changes in Ephemeroptera genera. \code{taxonFix = "remove"} prioritizes
accurate identifications of organisms by dropping observations from problematic
genera that do not have species-level identification. Without a species-level
identification, there is no way to assure correct membership in an updated genus.
Organisms with a species-level identification are cross-walked to an updated genus.
NOTE on “slash” genera: When \code{taxonFix = "lump"}, these "slash"
genera are rolled into the larger linked genera, as above. \code{taxonFix = "remove"}
prioritizes accurate identifications by dropping all slash genera are, including
those organisms identified as a “slash” genus at the bend; this option will
result in many fewer genera in the final dataset. Finally, \code{taxonFix = "none"},
includes "slash" genera, but it does not connect these genera to larger linked genera.
See \code{vignette("GettingStarted")} for more information.

\code{agency} refers to the \code{agency} that collected the invertebrate samples.
If you want to use data from both agencies, set \code{agency} = c("USGS", "EPA"),
which is the default. For USGS data, sampling data include all USGS BioData
with \code{SampleMethodCode} of "BERW", "IRTH", "SWAMP", "EMAP", "CDPHE", and "PNAMP".
Note that by default, only moving waters classified as "wadeable" are
included but setting \code{boatableStreams = TRUE} will include observations from
non-wadeable streams. Some information included in the EPA dataset are not
included in the USGS datasets, specifically observed wetted width of the
stream/river.

If \code{rarefy = TRUE}, only samples with 300+ individuals identified (raw
count) are retained. Thus, ~17 \\% of samples are removed, as they have <300
individuals sampled. The rarefaction threshold is 300 organisms for each
sampling event, because 1) with every 50 individuals identified, ~1 genera
are added to the sample and 2) 82.8 \\% of samples have at least 300 individuals
identified. Thus, lowering the threshold to 200 individuals removed ~2 genera
per sample, but only an additional 8.8 \\% of samples are included (90.1 \\% from 82.8 \\%).
Similarly, increasing the threshold to 400 individuals added ~2 genera per
sample, but reduced samples to 30.3 \\% of all samples. Use \code{seed = ...}
to get consistent output of community data. See \code{vignette("GettingStarted")}
for more information regarding rarefaction. NOTE: \code{rarefy = TRUE} can
be used when a user wants occurrence data (presence/absence) OR proportional
data (each taxon represents a certain proportion of a sample). Use
\code{rarefy = FALSE} when densities are the measure of interest.

When dataType = “abun”, the function calculates taxa densities from samples
using lab subsampling ratios and area sampled \deqn{Taxa~abundance = n * Subsampling~Ratio}
where \emph{n} is the number of specimens identified and \emph{Subsampling Ratio} is the
proportion of the sample that was identified at the lab bench. For
the USGS dataset, this incorporates both "field split ratio" (proportion of
the sample that was brought into the lab for specimen identification) and
the "lab subsampling ratio" (proportion of grids used to identify invertebrates
at the lab bench). For the EPA datasets, this is just the "lab subsampling ratio", the
proportion of grids used to identify invertebrates at the lab bench. Then,
\deqn{Taxa~density = Taxa~abundance / Area~sampled~(m^2)} where taxa abundance
is simply divided by area sampled to estimate density. See
\code{vignette("GettingStarted")} for more details on the calculation of
taxa densities.

Note: There are 81 sampling events (sampling location - collection
date) with replicate samples (not duplicates). We have left these replicate
samples in the dataset, because some replicates are different in the stream
habitat sampled, which may be of interest for certain ecological questions.
For users who would like to remove these data to reduce potential effects of
replicate samples on their analyses, use
\code{data \%>\% group_by(Site_Number, CollectionDayOfYear) \%>\% slice(1)} to retain
only one replicate from each sampling event from the dataset.
}
\examples{
\dontrun{
Inverts <- getInvertData(taxonLevel = "Genus")

RarefyInverts <- getInvertData(taxonLevel = "Genus",
                               rarefy = TRUE,
                               seed = 10)
}

}
\author{
Michael Mahon, Devin Jones, Samantha Rumschlag
}
