% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvertData.R
\name{getInvertData}
\alias{getInvertData}
\title{Access clean USGS Macro-Invertebrate Dataset}
\usage{
getInvertData(
  dataType = "occur",
  taxonLevel = "Genus",
  taxonFix = "lump",
  agency = c("USGS", "EPA"),
  lifestage = FALSE,
  rarefy = TRUE,
  sharedTaxa = FALSE,
  seed = 0
)
}
\arguments{
\item{dataType}{Output data type, either \code{"abun"} or \code{"occur"}.}

\item{taxonLevel}{Level of taxonomic resolution, must be one of:
\code{"Phylum"}, \code{"Class"}, \code{"Order"},
\code{"Family"}, \code{"Genus"}, or \code{"Species"}.}

\item{taxonFix}{How to deal with changes in taxonomy across time, must
be one of: \code{"none"}, \code{"lump"}, \code{"remove"}.}

\item{agency}{The agency name(s) (e.g., "USGS" and "EPA") that should be
included in the output dataset. As of now, "USGS" must be present in the
agency vector. See \code{Details} below for more information.}

\item{lifestage}{logical. Should the output dataset should include lifestage information for
each individual? \code{TRUE} or \code{FALSE}.}

\item{rarefy}{logical. Should samples be standardized by the number of individuals
identified within each sample? \code{TRUE} or \code{FALSE}. See \code{Details}
below for more information.}

\item{sharedTaxa}{logical. Should Genera be limited to those that appear in
both the NRSA and USGS datasets? \code{TRUE} or \code{FALSE}.}

\item{seed}{numeric. Set seed for \code{rarefy} to get consistent results with every
iteration of the function.}
}
\value{
A species by sample data frame with site, stream reach, and
  sample information.
}
\description{
Access clean USGS Macro-Invertebrate Dataset
}
\details{
Note: There are 81 sampling events (sampling location - collection
  date) with replicate samples (not duplicates). We have left these replicate
  samples in the dataset, because some replicates are different in the
  stream habitat sampled, which may be of interest for certain ecological
  questions.

  The function adjusts taxa abundances from samples for lab subsampling
  ratio. As a result, duplicate taxa were combined (abundances summed)
  following adjustments for lab subsampling.

  Currently, lab large rares are removed from all samples.

  \code{taxonFix} provides ways to handle changes in taxonomy across time,
  especially in instances in which species have been reorganized into new
  genera. \code{taxonFix} operates on the genera level. \code{taxonFix = "none"}
  makes no adjustment. \code{taxonFix = "lump"} prioritizes retaining observations
  by giving a unified genera name to all species and genera that have been linked
  through changes in taxonomy (e.g. genera1/genera2/genera3). Note: of 98
  problematic genera that exist throughout all BioData, \code{taxonFix = "lump"}
  results in 13 "lumped" genera. Most new "lumped" genera are <5 "old" genera joined.
  One "lumped" genera (within Ephemeroptera) includes 70 "old" genera.
  \code{taxonFix = "remove"} prioritizes accurate identification by dropping
  observations from problematic genera that do not have species identification.
  Without a species-level identification from the bench, there is no way to
  assure correct membership in a updated genera. NOTE: "slash" genera that NRSA
  and NAWQA roll up into family/subfamily are now being included as genera. As
  a result, there are fewer individual genera, but this results in more data
  being included at the genus level. When \code{taxonFix = "lump"}, these "slash"
  genera are rolled into the larger linked genera, as above. \code{taxonFix = "remove"}
  still prioritizes accurate identifications by dropping all obscure identifications,
  such that ALL slash genera are dropped; this will result in much less data
  being included in the final dataset. Finally, \code{taxonFix = "none"}, still
  generate "slash" genera, but, it does not link these genera to larger linked
  genera.

  \code{agency} refers to agency that collected the invertebrate samples. If
  you want to use data from both agencies, set \code{agency = c("USGS", "EPA")}.
  For the "USGS" dataset, this includes all programs with SampleMethodCodes
  of "BERW", "IRTH", "SWAMP", "EMAP", "CDPHE", and "PNAMP". If \code{agency}
  includes "EPA", samples from the EPA National Stream and River
  Assessment programs (2013-2014, 2008-2009) and Wadeable Stream Assessment
  (2000-2004) will be included. Note that from these samples, only moving
  waters classified as "wadeable" are included and only samples that are
  "reach-wide" are included. Some information included in the USGS dataset
  are not included in the EPA datasets, and thus will appear as "NA".
  Similar to the USGS data, there were inherent taxonomic issues with the
  EPA data. As such, we have taken the same steps as described above under
  \code{taxonFix} to address these concerns. NOTE: As of now, \code{agency}
  must include "USGS". This will be fixed in future versions.

  If \code{rarefy = TRUE}, only samples with 300+ individuals identified (RawCount)
  will be retained. Thus, ~17 \% of samples will be removed, as they have <300
  individuals sampled. We set the rarefaction threshold at 300, because 1) with
  every 50 individuals identified, ~1 genera are added to the sample and 2) 82.8 \%
  of samples have at least 300 individuals identified. Thus, lowering the threshold
  to 200 individuals removed ~2 genera per sample, but only added an additional
  7.3 \% of samples included (90.1 \% from 82.8 \%). Similarly, increasing the
  threshold to 400 individuals added ~2 genera per sample, but reduced samples to
  30.3 \% of all samples. Use \code{seed = ...} to get consistent output of
  community data. NOTE: \code{rarefy = TRUE} can be used when wanting
  occurrence data (presence/absence) OR proportional data (each taxon represents
  a certain percent of a sample). Use \code{rarefy = FALSE} when densities are
  the measure that you are interested in using.
}
\examples{
\dontrun{
Inverts <- getInvertData(taxonLevel = "Genus")

RarefyInverts <- getInvertData(taxonLevel = "Genus",
                               rarefy = TRUE,
                               seed = 10)
}

}
