% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvertData.R
\name{getInvertData}
\alias{getInvertData}
\title{Access and harmonize macroinvertebrate data}
\usage{
getInvertData(
  dataType = "occur",
  taxonLevel = "Genus",
  taxonFix = "lump",
  agency = c("USGS", "EPA"),
  lifestage = FALSE,
  rarefy = TRUE,
  rarefyCount = 300,
  sharedTaxa = FALSE,
  seed = 0,
  boatableStreams = FALSE
)
}
\arguments{
\item{dataType}{Output data type for the community matrix, either \code{"density"}
(density) or \code{"occur"} (occurrence).}

\item{taxonLevel}{Level of taxonomic resolution for the community matrix. Input must be one of:
\code{"Phylum"}, \code{"Class"}, \code{"Order"},
\code{"Family"}, \code{"Subfamily"}, \code{"Genus"}, or \code{"Mixed"}.}

\item{taxonFix}{Option to account for changes in taxonomy across time, must be
one of: \code{"none"}, \code{"lump"}, \code{"remove"}. See \code{Details}
below for more information.}

\item{agency}{The agency name or names (e.g., "USGS" and "EPA") that are the
source of data for the output community matrix. See \code{Details} below for more information.}

\item{lifestage}{logical. For USGS data only, should the output dataset include
lifestage information for each taxa?  \code{TRUE} or \code{FALSE}. Default
is \code{FALSE}.}

\item{rarefy}{logical. Should samples be standardized by the number of individuals
identified within each sample? \code{TRUE} or \code{FALSE}. See \code{Details}
below for more information.}

\item{rarefyCount}{integer. If \code{rarefy = TRUE}, the individual count to be used
as the cutoff for rarefaction (standardizing samples by the number of individuals
identified). Default is 300.}

\item{sharedTaxa}{logical. Should Genera be limited to those that appear in
both the EPA and USGS datasets? \code{TRUE} or \code{FALSE}. Must be set to
\code{FALSE} when only one agency is specified. Default is \code{FALSE}}

\item{seed}{numeric. Set seed for \code{rarefy} to get consistent results with each new
iteration of the function. Value gets passed to \code{set.seed} internally.}

\item{boatableStreams}{logical. Should EPA boatable streams be included in the
output dataset?  \code{TRUE} or \code{FALSE}. Note: all USGS streams are wadable.
It is not advisable to include boatable streams when building a dataset including
both EPA and USGS data. Boatable EPA data and wadeable USGS data are not
considered comparable.}
}
\value{
A taxa by sample data frame with site, stream reach, and sample information.
}
\description{
This function generates an occurrence or abundance community matrix for
benthic macroinvertebrates sampled in rivers and streams from the US EPA
National Rivers and Streams Assessment and USGS BioData.
}
\details{
\code{agency} refers to the \code{agency} that collected the invertebrate samples.
If you want to use data from both agencies, set \code{agency} = c("USGS", "EPA"),
which is the default. For USGS data, sampling data include all USGS BioData
with \code{SampleMethodCode} of "BERW", "IRTH", "SWAMP", "EMAP", "CDPHE RR", and "PNAMP".
Note that by default, only moving waters classified as "wadeable" are
included but setting \code{boatableStreams = TRUE} will include observations from
non-wadeable streams. Some information included in the EPA dataset are not
included in the USGS datasets, specifically observed wetted width of the
stream/river.

\code{taxonLevel} refers to the taxonomic resolution (Genus, Class, Family, etc.)
for the sample by taxa matrix. The input values for this parameter are case
sensitive and must start with a capital letter. All observations taxonomically
coarser than the \code{taxonLevel} provided are dropped from the output community matrix.
For instance, if \code{taxonLevel = "Genus"} , then observations identified at
Subfamily, Family, Order, Class, or Phylum levels are dropped. When
\code{taxonLevel = "Subfamily"}, for taxa without subfamilies, the Family-level
designation is returned. When \code{taxonLevel = "Subfamily"}, for taxa without subfamilies, the Family-level
designation is returned. "Genus" is the finest level of taxonomic resolution
provided for macroinvertebrates. The function also provides the option of returning
the lowest level of taxonomic identification for all specimen (\code{taxonLevel = "Mixed"}).

\code{taxonFix} provides options to account for changes in taxonomy across time,
especially in instances in which species have been reorganized into new genera.
\code{taxonFix} operates on the genera level. \code{taxonFix = "none"}
makes no adjustment. \code{taxonFix = "lump"} prioritizes retaining observations
by giving a unified "slash" genera name to all species and genera that have been linked
through changes in taxonomy through time (e.g. genera1/genera2/genera3). Note:
of 82 problematic genera that exist throughout both datasets,
\code{taxonFix = "lump"} results in 11 "lumped" genera. All but two complexes
of genera were composed of two individual genera. A single complex of genera
within the Ephemeroptera order included 70 individual genera, and a second
complex of Ephemeroptera genera included six individual genera. Because of
the complexity of Ephemeroptera taxonomy, careful consideration should be
given to inferences that can be made when evaluating Ephemeroptera trends.
The authors, for instance, do not advise that users evaluate temporal changes
in abundance or richness of the linked groups of Ephemeroptera genera because
taxonomic reorganizations likely obscure temporal patterns.

Alternatively, genera linked by taxonomic reorganization can be removed with
\code{taxonFix = "remove"}. This option prioritizes accurate identification by
dropping observations that cannot be confidently identified to a single genus,
as in the complexes of genera previously described.  Without a species-level
identification, there is no way to assure correct membership in an updated genus.
Organisms with a species-level identification are cross-walked to an updated genus.
NOTE on “slash” genera: When \code{taxonFix = "lump"}, these "slash"
genera are rolled into the larger linked genera, as above. \code{taxonFix = "remove"}
prioritizes accurate identifications by dropping all slash genera are, including
those organisms identified as a “slash” genus in the lab; this option will
result in many fewer genera in the final dataset. Finally, \code{taxonFix = "none"},
includes "slash" genera, but it does not connect these genera to larger linked genera.

\code{taxonFix} operates on the genus level, so set \code{taxonFix = "none"}
when taxonLevel is set to \code{"Family"} or higher taxonomic resolution.
Care should be taken to harmonize taxonomy either with the approaches provided
or some alternative when long time scale datasets on the entire community
of macroinvertebrates are generated because changes in taxonomy can make it
artificially appear as though some genera are either appearing or disappearing
in time. See \code{vignette("GettingStarted")} for more information.

If \code{rarefy = TRUE}, samples with \code{rarefyCount}+ individuals
identified (raw count) are retained. Thus, a percentage of samples will be
removed, as they have <\code{rarefyCount} individuals sampled. The rarefaction
threshold is default is 300 organisms, because 1) with every 50 individuals
identified, ~1 genera are added to the sample and 2) 91.3 \\% of samples have
at least 300 individuals identified. Thus, lowering the threshold to 200
individuals removed ~2 genera per sample, but only an additional 3.2 \\% of
samples are included (94.5 \\% from 91.3 \\%). Similarly, increasing the threshold
to 400 individuals added ~2 genera per sample, but reduced samples to 70.6 \\% of
all samples. Use \code{seed = ...}  to get consistent output of community data.
See \code{vignette("GettingStarted")} for more information regarding
rarefaction. NOTE: \code{rarefy = TRUE} can be used when a user wants
occurrence data (presence/absence) OR proportional data (each taxon represents
a certain proportion of a sample). Use \code{rarefy = FALSE} when densities
are the measure of interest.

When dataType = "density", the function calculates taxa densities from samples
using lab subsampling ratios and area sampled
\deqn{Taxa~abundance = n*frac{1}{PropID}}
\deqn{Taxa~density = frac{Taxa~abundance}{Area~sampled~(m^2)}}
where \emph{n} is the number of specimens identified and \emph{PropID} is the
proportion of the sample that was identified at the lab bench. For
the USGS dataset, this incorporates both "field split ratio" (proportion of
the sample that was brought into the lab for specimen identification) and
the "lab subsampling ratio" (proportion of grids used to identify invertebrates
at the lab bench). For the EPA datasets, this is just the "lab subsampling ratio", the
proportion of grids used to identify invertebrates at the lab bench. See
\code{vignette("GettingStarted")} for more details on the calculation of
taxa densities.
}
\examples{
\dontrun{
Inverts <- getInvertData(taxonLevel = "Genus")

RarefyInverts <- getInvertData(taxonLevel = "Genus",
                               rarefy = TRUE,
                               seed = 10)
}

}
\author{
Michael Mahon, Devin Jones, Samantha Rumschlag, Terry Brown
}
