% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNLCDData.R
\name{getNLCDData}
\alias{getNLCDData}
\title{Generate NLCD Land-use/Land-cover data for each NAWQA site}
\usage{
getNLCDData(data, scale = "cat", group = FALSE)
}
\arguments{
\item{data}{Input dataset which needs to include columns for
\code{"SiteNumber"} and \code{"CollectionYear"}.
\code{"SiteNumber"} must have \code{"USGS-"} prefix.}

\item{scale}{Scale for LULC data to be generated. Must be \code{"cat"} for
catchment or \code{"ws"} for watershed. See \code{Details} below for more
information.}

\item{group}{logical. Should the land-use/land-cover data be grouped into broad
categories? \code{TRUE} or \code{FALSE}. See \code{Details} below for more
information.}
}
\value{
Function returns a site by year dataframe of NLCD land-use/land-cover
  data for either the catchment or watershed scale.
}
\description{
Generate NLCD Land-use/Land-cover data for each NAWQA site
}
\details{
To be completed.
}
\examples{
\dontrun{
## Code to generate the sum of the county-level use estimates (high estimate)
## from site number "05276005" from 1997-1999 (2 year lag) for "insecticide",
## "herbicide", and "synthetic_fungicide".

  dat = data.frame(SiteNumber = "USGS-05276005",
                   CollectionYear = 2007)

  getNLCDData(data=dat,
  scale = "cat",
  group = FALSE
  }

}
