% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumpRollUp.R
\name{lumpRollUp}
\alias{lumpRollUp}
\title{Roll 'lumped' genera to lowest common taxonomic resolution}
\usage{
lumpRollUp(inputdataset)
}
\arguments{
\item{inputdataset}{Input dataset that must be generated from
\code{getInvertData()}.}
}
\value{
A taxa by sample data frame with site, stream reach, and sample information.
}
\description{
This function replaces all 'lump' genera (generated by
\code{getInvertData(taxonFix = "lump")}) with the lowest common taxonomic
resolution for all genera in the lumped genus. Designed to be implemented
with \code{getInvertData(taxonLevel = "Mixed")}.
}
\details{
In some instances, like when calculating MMIs for example, it may be useful
to roll the 11 "lumped" genera to their lowest common taxonomic resolution
(Subfamily, Family, or Order). This function does not adjust "slash" taxa, or
those organisms that were identified to a slash genus designation in the lab
(see the Getting Started vignette for more details). This function gathers the lowest common
taxonomic resolution all lump genus members share. The function then replaces the "lump genus" name
(e.g., Elimia/Pleurocera) with the lowest common taxonomic resolution
designation (e.g., Pleuroceridae). The function then sums occurrence or density data
within duplicate broader taxonomic designations. For instance, if
\code{taxonLevel = "Mixed"}, then observations among duplicated taxonomic unit
columns would be summed (occurrences summed then converted to 0/1).

The input dataset should be directly generated by \code{getInvertData()}, with
\code{taxonFix = "lump"}.
}
\examples{
\dontrun{
Inverts <- getInvertData(taxonLevel = "Mixed",
                         taxonFix = "lump")

##roll lumped genera to lowest common taxonomic unit
InvertsUpdatedLump <- lumpRollUp(Inverts)

}

}
\author{
Michael Mahon
}
